refactor: create components to avoid using 'use client' directive on login and register pages
